// frontend/src/api/models/handlerSalesSummaryResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Sales summary data for a period
 */
export interface HandlerSalesSummaryResponse {
  avg_order_value?: number
  period_end?: string
  period_start?: string
  profit_margin?: number
  total_cost_amount?: number
  total_gross_profit?: number
  total_orders?: number
  total_quantity?: number
  total_sales_amount?: number
}