// frontend/src/api/models/handlerProductSalesRankingResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Product sales ranking item
 */
export interface HandlerProductSalesRankingResponse {
  category_name?: string
  order_count?: number
  product_id?: string
  product_name?: string
  product_sku?: string
  rank?: number
  total_amount?: number
  total_profit?: number
  total_quantity?: number
}