// frontend/src/api/models/handlerCustomerSalesRankingResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Customer sales ranking item
 */
export interface HandlerCustomerSalesRankingResponse {
  customer_id?: string
  customer_name?: string
  rank?: number
  total_amount?: number
  total_orders?: number
  total_profit?: number
  total_quantity?: number
}